refactor fft planning
